+++ DatadogInternal/Sources/Context/DatadogContext.swift
@@ -71,6 +71,9 @@ public struct DatadogContext {
     /// Current device information.
     public var device: DeviceInfo
 
+    /// Current locale information.
+    public var localeInfo: LocaleInfo
+
     /// Operating System information.
     public let os: OperatingSystem
 
@@ -110,6 +113,9 @@ public struct DatadogContext {
     /// This value can be `nil` of the current device battery interface is not available.
     public var batteryStatus: BatteryStatus?
 
+    /// The current brightness status.
+    public var brightnessLevel: BrightnessLevel?
+
     /// `true` if the Low Power Mode is enabled.
     public var isLowPowerModeEnabled = false
 
@@ -135,6 +141,7 @@ public struct DatadogContext {
         applicationBundleType: BundleType,
         sdkInitDate: Date,
         device: DeviceInfo,
+        localeInfo: LocaleInfo,
         os: OperatingSystem,
         nativeSourceOverride: String? = nil,
         userInfo: UserInfo? = nil,
@@ -145,6 +152,7 @@ public struct DatadogContext {
         networkConnectionInfo: NetworkConnectionInfo? = nil,
         carrierInfo: CarrierInfo? = nil,
         batteryStatus: BatteryStatus? = nil,
+        brightnessLevel: BrightnessLevel? = nil,
         isLowPowerModeEnabled: Bool = false,
         additionalContext: [String: AdditionalContext] = [:]
     ) {
@@ -165,6 +173,7 @@ public struct DatadogContext {
         self.applicationBundleType = applicationBundleType
         self.sdkInitDate = sdkInitDate
         self.device = device
+        self.localeInfo = localeInfo
         self.os = os
         self.nativeSourceOverride = nativeSourceOverride
         self.userInfo = userInfo
@@ -175,6 +184,7 @@ public struct DatadogContext {
         self.networkConnectionInfo = networkConnectionInfo
         self.carrierInfo = carrierInfo
         self.batteryStatus = batteryStatus
+        self.brightnessLevel = brightnessLevel
         self.isLowPowerModeEnabled = isLowPowerModeEnabled
         self.additionalContext = additionalContext
     }