+++ DatadogRUM/Sources/Instrumentation/Views/SwiftUI/SwiftUIRUMViewsPredicate+objc.swift
@@ -62,17 +62,14 @@ public protocol objc_SwiftUIRUMActionsPredicate: AnyObject {
 public class objc_DefaultSwiftUIRUMActionsPredicate: NSObject, objc_SwiftUIRUMActionsPredicate {
     private let swiftPredicate: DefaultSwiftUIRUMActionsPredicate
 
+    @objc(initWithIsLegacyDetectionEnabled:)
     public init(isLegacyDetectionEnabled: Bool) {
         swiftPredicate = DefaultSwiftUIRUMActionsPredicate(isLegacyDetectionEnabled: isLegacyDetectionEnabled)
         super.init()
     }
 
-    override public convenience init() {
-        self.init(isLegacyDetectionEnabled: true)
-    }
-
     public func rumAction(with componentName: String) -> objc_RUMAction? {
-        swiftPredicate.rumAction(with: componentName).map {
+        return swiftPredicate.rumAction(with: componentName).map {
             objc_RUMAction(name: $0.name, attributes: $0.attributes.dd.objCAttributes)
         }
     }